chore: move build-plugins under config/
Keep the root free of build tooling by colocating Vite/electron-vite
plugins with the rest of config. The plain-Node entry guard and
bootstrap fatal-exit banner behavior are unchanged; only paths move.
