+++ .github/workflows/win-crash-survival-e2e.yml
@@ -77,7 +77,6 @@ jobs:
               '!config/reliability-gates.jsonc',
               '!config/max-lines-baseline.txt',
               '!config/vitest.config.ts',
-              'build-plugins/**',
               'native/**',
               'resources/**',
               'electron.vite.config.ts',
