+++ config/build-plugins/plain-node-entry-guard.ts
