+++ config/tsconfig.node.json
@@ -2,7 +2,7 @@
   "extends": "@electron-toolkit/tsconfig/tsconfig.node.json",
   "include": [
     "../electron.vite.config.*",
-    "../build-plugins/**/*",
+    "./build-plugins/**/*",
     "../src/main/**/*",
     "../src/renderer/src/lib/skill-freshness-display-status.ts",
     "../src/preload/**/*",
